Se agregó el msg de confirmación para Horario y Materia

+++ PdmSubeCum/app/src/main/java/com/pdmsubecum/mm14031/HorarioEliminarActivity.java
@@ -1,8 +1,11 @@
 package com.pdmsubecum.mm14031;
 
 import android.app.Activity;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
@@ -21,6 +24,7 @@ public class HorarioEliminarActivity extends Activity {
         setContentView(R.layout.activity_horario_eliminar);
         helper = new DataBase(this);
         editIdHorario = (EditText) findViewById(R.id.editIdHorario);
+
     }
 
     public boolean verificarDatos(){
@@ -37,14 +41,59 @@ public class HorarioEliminarActivity extends Activity {
 
         if(verificarDatos()){
             String mensaje ;
-            Horario horario = new Horario();
-            horario.setIdHorario(Integer.parseInt(editIdHorario.getText().toString()));
+            boolean grupoRelacionExiste;
+            int idHorario;
+            idHorario =Integer.parseInt(editIdHorario.getText().toString());
 
             helper.abrir();
-            mensaje = helper.eliminarHorario(horario);
+            grupoRelacionExiste = helper.grupoRelacionExiste(idHorario);
             helper.cerrar();
 
-            Toast.makeText(this,mensaje,Toast.LENGTH_SHORT).show();
+
+            if(grupoRelacionExiste){
+                AlertDialog.Builder dialogo = new AlertDialog.Builder(HorarioEliminarActivity.this);
+                dialogo.setIcon(R.mipmap.ic_launcher).
+                        setMessage("Importante").
+                        setMessage("El Horario está relacionada con otros registros.\n\n ¿Desea eliminar en cascada?").
+                        setPositiveButton("Aceptar", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                eliminar();
+                                Toast.makeText(HorarioEliminarActivity.this, "Eliminado Satisfactoriamente", Toast.LENGTH_SHORT).show();
+                            }
+                        }).setNegativeButton("Cancelar", new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                    }
+                });
+                dialogo.create();
+                dialogo.show();
+            }else{
+                Horario horario = new Horario();
+                horario.setIdHorario(idHorario);
+
+                helper.abrir();
+                mensaje = helper.eliminarHorario(horario);
+                helper.cerrar();
+
+                Toast.makeText(this,mensaje,Toast.LENGTH_SHORT).show();
+            }
         }
     }
+
+
+    public void eliminar(){
+        String mensaje ;
+        Horario horario = new Horario();
+        horario.setIdHorario(Integer.parseInt(editIdHorario.getText().toString()));
+
+        helper.abrir();
+        mensaje = helper.eliminarHorario(horario);
+        helper.cerrar();
+
+        Toast.makeText(this,mensaje,Toast.LENGTH_SHORT).show();
+    }
+
+
 }
